src/phase1/requirements.txt
azure-ai-projects
azure-ai-inference[prompts]
azure-core
azure-identity
azure-monitor-opentelemetry
azure-search-documents
pandas
python-dotenv
opentelemetry-api